lkl: integrate with zpoline for alternate hijack backend
This commit introduces an integration with zpoline (*1), which is a
mechanism to rewrite binary upon loading.  zpoline allows us to replace
symbols of system call to different functions, which current LKL hijack
library does the same thing in a different way, by using symbol
replacement via LD_PRELOAD.

The benefit of zpoline is that the replacement of syscall is at the
instruction of `syscall` or `sysenter`, which userspace program can
catch all syscalls,  while the approach based on LD_PRELOAD cannot when
the symbols of interet are hidden within libc (e.g., __socket).

For more detail about the internal of zpoline, take a look at *1.

*1: https://github.com/yasukata/zpoline

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>